requirements.txt
Django > 3.1.4
djangorestframework
djangorestframework-simplejwt
django-background-tasks
httpx
requests